Fixed typo in RoundEnvironementInternal .
