+++ specs/src/main/java/io/nuun/kernel/api/plugin/RoundEnvironmentInternal.java
@@ -1,0 +1,54 @@
+/**
+ * Copyright (C) 2013-2014 Kametic <epo.jemba@kametic.com>
+ *
+ * Licensed under the GNU LESSER GENERAL PUBLIC LICENSE, Version 3, 29 June 2007;
+ * or any later version
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.gnu.org/licenses/lgpl-3.0.txt
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.nuun.kernel.api.plugin;
+
+/**
+ * Default implementation of {@link io.nuun.kernel.api.plugin.RoundEnvironment}.
+ */
+public class RoundEnvironmentInternal implements RoundEnvironment
+{
+
+    private int roundNumber = 0;
+
+    /**
+     * Constructor.
+     */
+    public RoundEnvironmentInternal()
+    {
+    }    
+    
+    @Override
+    public int roundNumber()
+    {
+        return this.roundNumber;
+    }
+
+    /**
+     * Increments the round number.
+     */
+    public void incrementRoundNumber()
+    {
+        this.roundNumber++;
+    }
+
+    @Override
+    public boolean firstRound()
+    {
+        return this.roundNumber == 0;
+    }
+
+}
